+++ graphannis/src/annis/db/aql/operators/edge_op.rs
@@ -302,7 +302,7 @@ impl BinaryOperatorBase for BaseEdgeOp {
                 }
                 // get number of nodes reachable from min to max distance
                 let max_dist = match self.spec.dist.max_dist() {
-                    std::ops::Bound::Unbounded => usize::max_value(),
+                    std::ops::Bound::Unbounded => usize::MAX,
                     std::ops::Bound::Included(max_dist) => max_dist,
                     std::ops::Bound::Excluded(max_dist) => max_dist - 1,
                 };